requirements.txt
streamlit
Pillow
numpy
torch
torchvision
matplotlib
pytorch_lightning
opencv-python-headless
timm
